BasePhyKilosort extractor
